popup/view/MnemonicPhrase: Reskin ConfirmMnemonicPhrase (#359)
* first cut with todos

* use copyText and button from sds

* use scss

* cleanup

* cleanup

* reskin confirm mnemonic phrase

* BEM css naming

* checkbutton to css
