+++ extension/src/popup/ducks/accountServices.ts
@@ -156,8 +156,7 @@ export const confirmMnemonicPhrase = createAsyncThunk<
     } else {
       return thunkApi.rejectWithValue({
         applicationState: res.applicationState,
-        errorMessage:
-          "The secret phrase you entered is invalid, please check the phrase you have noted and try again.",
+        errorMessage: "The secret phrase you entered is incorrect.",
       });
     }
